+++ examples/reference_tool.py
@@ -1,6 +1,7 @@
 """
 A tool to compare randomly generated COBEA results of different runs and versions.
-Note: Does not work well for M=2 yet, possibly due to not enough contraints in model_generator.
+Note: Does not work well for M=2 from model_generator yet, but does so for real responses
+(possibly not enough contraints in model_generator for this case).
 
 Bernard Riemann (bernard.riemann@tu-dortmund.de)
 """
@@ -15,8 +16,11 @@ reference_dict = {'set1': (30,32,1,0.02)} # (K, J, M, relative_noise)
 def compare_results(new_result, old_result):
     for attribute in ('R_jmw', 'A_km', 'mu_m', 'd_jw', 'b_k'):
         arr = [getattr(result,attribute) for result in (new_result, old_result)]
-        assert_array_almost_equal_nulp(arr[0], arr[1], 3)
-    print('%s: present result is approx. equivalent to reference.' % reference_name)
+        # try:
+        assert_array_almost_equal_nulp(arr[0], arr[1])
+        # except AssertionError:
+        #     print('%s is different: computation %s vs reference %s' % (attribute, arr[0].repr(), arr[1].repr))
+    print('present result is approx. equivalent to reference.')
 
     for key in ('coretime',):
         try:
@@ -25,43 +29,62 @@ def compare_results(new_result, old_result):
             pass
 
 
+def compare_computation_to_reference(response_filename, result_filename, make_reference=False):
+
+    with open(response_filename,'rb') as f:
+         response = load(f)
+
+    result = cobea(response)
+
+    if make_reference:
+        result.save(result_filename)
+    else:
+        with open(result_filename,'rb') as f:
+            reference_result = load(f)
+        compare_results(result, reference_result)
+
+
 if __name__=='__main__':
     from cobea import cobea
+    from sys import argv
 
-    overwrite = False
-    for reference_name in reference_dict:
-
-        response_filename = reference_name+'/response_input.pickle'
-        result_filename = reference_name+'/cobea_result.pickle'
-
+    if len(argv) > 2:
         try:
-            makedirs(reference_name)
-        except OSError:
-            pass # directory already exists
+            compare_computation_to_reference(argv[1], argv[2], argv[3] == 'ref')
+        except IndexError:
+            compare_computation_to_reference(argv[1], argv[2])
 
-        if overwrite:
-            K, J, M, relative_noise = reference_dict[reference_name]
-            question = 'x'
-            while question not in ('Y', 'N'):
-                question = input('really overwrite %s? (please type Y or N)' % reference_name)
-            if question == 'Y':
-                response, drift_space = random_response_drift(K, J, M, make_drift=False,
-                                                              hidden_filename=reference_name+'/hidden_model.pickle')
-                response.save(response_filename)
+    else:
+        overwrite = False
+
+        for reference_name in reference_dict:
+            print('-------- %s --------' % reference_name)
+
+            try:
+                makedirs(reference_name)
+            except OSError:
+                pass # directory already exists
+
+
+            response_filename = reference_name+'/response_input.pickle'
+            result_filename = reference_name+'/cobea_result.pickle'
+
+            if overwrite:
+                K, J, M, relative_noise = reference_dict[reference_name]
+                question = 'x'
+                while question not in ('Y', 'N'):
+                    question = input('really overwrite %s? (please type Y or N)' % reference_name)
+                if question == 'Y':
+                    response, drift_space = random_response_drift(K, J, M, make_drift=False,
+                                                                  hidden_filename=reference_name+'/hidden_model.pickle')
+                    response.save(response_filename)
+                else:
+                    continue
+
+                result = cobea(response)
+
+                result.save(result_filename)
+                print('%s: new reference set saved / overwritten' % reference_name)
+
             else:
-                continue
-        else:
-            with open(response_filename,'rb') as f:
-                response = load(f)
-
-        result = cobea(response)
-
-        if overwrite:
-            result.save(result_filename)
-            print('%s: new reference set saved / overwritten' % reference_name)
-        else:
-            with open(result_filename,'rb') as f:
-                reference_result = load(f)
-            compare_results(result, reference_result)
-        print('---')
-
+                compare_computation_to_reference(response_filename, result_filename)